refactor: renamed classes
